refactor: vip-453 to vip-454
